requirements.txt
requests==2.28.1
pytest==9.0.1
python-dotenv==1.2.1
mysql-connector-python==9.5.0
pytest-sugar==1.1.1